Added action LeaveFromGroup, Extend some actions from Invitation and Group Controller about authorizing mechanism

+++ MyProjectRestApi/Controllers/InvitationsController.cs
@@ -6,6 +6,7 @@ using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
@@ -33,9 +34,10 @@ namespace MyProjectRestApi.Controllers
         }
 
         // GET: api/Invitations?userId=b3913d15-f95f-412f-b395-a11d281df016(this is userid)
-        public IQueryable<InvitationDTO> GetInvitationsByUserId(string userId)
+        public IQueryable<InvitationDTO> GetInvitationsByUserId()
         {
-            return db.Invitations.Where(userid => userid.ApplicationUser.Id == userId)
+            string currentUserId = GetCurrentUserId();
+            return db.Invitations.Where(userid => userid.ApplicationUser.Id == currentUserId)
                 .Select(invite => new InvitationDTO()
                 {
                     Id = invite.Id,
@@ -110,38 +112,55 @@ namespace MyProjectRestApi.Controllers
         }
 
         // POST: api/Invitations
-        [ResponseType(typeof(InvitationDTO))]
-        public async Task<IHttpActionResult> PostInvitation(Invitation invitation)
+        //[ResponseType(typeof(List<InvitationDTO>))]
+        public async Task<IHttpActionResult> PostInvitation(List<InvitationDTO> invitationsDTO)
         {
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            ApplicationUser userReceiverInvitaton = db.Users.Find(invitation.ApplicationUser.Id);
-            if (userReceiverInvitaton == null)
-                return BadRequest();
+            List<Invitation> invitations = new List<Invitation>();
+            foreach(InvitationDTO invitationDto in invitationsDTO)
+            {
+                ApplicationUser userReceiverInvitaton = db.Users.Find(invitationDto.UserIdReceiver);
+                if (userReceiverInvitaton == null)
+                    return BadRequest();
+                try
+                {
+                    Invitation hasInvitationAlreadyExist = db.Invitations
+                                                        .Where(invite => invite.ApplicationUser.Id == invitationDto.UserIdReceiver
+                                                        && invite.GroupId == invitationDto.GroupId)
+                                                        .FirstOrDefault();
 
-            Invitation hasInvitationAlreadyExist = await db.Invitations
-                                                        .Where(invite => invite.ApplicationUser.Id == invitation.ApplicationUser.Id
-                                                        && invite.GroupId == invitation.GroupId)
-                                                        .FirstOrDefaultAsync();
-
-            int hasUserAlreadyExistInGroup = (from gr in db.Groups
-                                              where (gr.Id == invitation.GroupId)
-                                              let userInGroup = gr.Users.Where(m => m.Id == invitation.ApplicationUser.Id)
-                                              where userInGroup.Any()
-                                              select userInGroup.Count()).FirstOrDefault();
+                    int hasUserAlreadyExistInGroup = (from gr in db.Groups
+                                                      where (gr.Id == invitationDto.GroupId)
+                                                      let userInGroup = gr.Users.Where(m => m.Id == invitationDto.UserIdReceiver)
+                                                      where userInGroup.Any()
+                                                      select userInGroup.Count()).FirstOrDefault();
 
 
-            if (hasInvitationAlreadyExist != null || hasUserAlreadyExistInGroup != 0)
-                return BadRequest();
+                    if (hasInvitationAlreadyExist != null || hasUserAlreadyExistInGroup != 0)
+                        return Content(HttpStatusCode.PreconditionFailed, "User already exist in the Group");
+                    
+                }catch(Exception ex)
+                {
+                    return Content(HttpStatusCode.InternalServerError, ex.Message);
+                }
+            
 
-            invitation.ApplicationUser = userReceiverInvitaton;
-            db.Invitations.Add(invitation);
+                Invitation inviteToGroup = new Invitation(invitationDto, userReceiverInvitaton);
+                invitations.Add(inviteToGroup);
+            }
+
+
+            invitations.ForEach(invite => db.Invitations.Add(invite));
             await db.SaveChangesAsync();
 
-            return CreatedAtRoute("DefaultApi", new { id = invitation.Id }, new InvitationDTO() { GroupId = invitation.GroupId, Id = invitation.Id });
+            //I can not return any value because Invitation is send to frends and I have not to get any invitation back
+            return StatusCode(HttpStatusCode.NoContent);
+            //return Ok();
+           // return CreatedAtRoute("DefaultApi", new { id = invitation.Id }, new InvitationDTO() { GroupId = invitation.GroupId, Id = invitation.Id });
         }
 
         // DELETE: api/Invitations/5
@@ -153,11 +172,14 @@ namespace MyProjectRestApi.Controllers
             {
                 return NotFound();
             }
+            string currentUserId = GetCurrentUserId();
+            if (!IsUserHasThisInvite(id, currentUserId))
+                return BadRequest();
 
             db.Invitations.Remove(invitation);
             await db.SaveChangesAsync();
 
-            return Ok(invitation);
+            return Ok();
         }
 
         protected override void Dispose(bool disposing)
@@ -173,5 +195,26 @@ namespace MyProjectRestApi.Controllers
         {
             return db.Invitations.Count(e => e.Id == id) > 0;
         }
+
+        private string GetCurrentUserId()
+        {
+            var identity = User.Identity as ClaimsIdentity;
+            Claim identityClaim = identity.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier);
+
+            var user = db.Users.FirstOrDefault(u => u.Id == identityClaim.Value);
+
+            return user.Id;
+        }
+
+        private bool IsUserHasThisInvite(int GroupId, string UserIdReceiver)
+        {
+            Invitation isUserHasThisInvite = db.Invitations
+                                               .Where(invite => invite.ApplicationUser.Id == UserIdReceiver
+                                               && invite.GroupId == GroupId)
+                                               .FirstOrDefault();
+            return (isUserHasThisInvite != null);
+
+        }
+
     }
 }
